Rename files to .js so we can simplify component imports without extensions
