+++ ui/src/client/app/Logs.js
