+++ ui/webpack.config.js
@@ -5,7 +5,7 @@ var BUILD_DIR = path.resolve(__dirname, 'src/client/public');
 var APP_DIR = path.resolve(__dirname, 'src/client/app');
 
 var config = {
-    entry: APP_DIR + '/index.jsx',
+    entry: APP_DIR + '/App.js',
     output: {
         path: BUILD_DIR,
         filename: 'bundle.js'
